fixed bug: Crash on delete wargame
